move prepare_for_draft to EagleDraftWorkerBase
